requirements.txt
discord.py
asyncpg
jishaku
git+https://github.com/bijij/discord-ext-store-true
